Modificaitons to the overview
I think we should leave off details about things like idempotency from
the overview and focus on just laying the ground work for what it is.
Such things can be discussed in more advanced sections.

Also we decided to refer to the new agent as dcm-agent (and not rely
on the implmentation detail of the language it is written in for
description).  Also the Java agent is referred to as the legacy agent.